refactor: muda tipo do value para FLOAT em todas as tabelas que contem essa coluna

+++ src/database/migrations/20220716004858-table-customerStockTransaction.js
@@ -25,7 +25,7 @@ module.exports = {
       },
       value: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT(20, 2)
       },
       quantity: {
         allowNull: false,
